requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
PyYAML==6.0
scipy==1.10.1
seaborn==0.12.2
tqdm==4.66.2
